Don't try to bump versions of excluded packages
Suppose a workspace containing:

   * Crate A, v1
   * Crate B, v2

A command like `cargo release -p crate-a 1.1.0` would fail because
`cargo release` would attempt to bump crate B to v1.1.0 while loading it
(which is not legal). This goes unnoticed when running something like
`cargo release -p crate-a minor` because crate B's version gets bumped
to v2.1.0 (and then this in-memory version bump is harmlessly
discarded).

The solution here is not the most elegant, but it works.